Add GetPackContents() logic
This was a huge leap in terms of SQL understanding. Referencing the id in a table given another column value in that table to select from a second table.
It took considerable time to parse out the syntax and felt really good when it returned a list.

Still need to flesh out CreateCard and DeleteCard to round off the PackManager but doing so will also make significant headway into the CardManager.