Add new models and data context
Implemented new models for Flashcard, Stack and StudySession in the 'Models' namespace. A new DbContext named 'FlashcardContext' was also created under the 'Data_Access' namespace, which includes DbSets for these three models and overrides the 'OnConfiguring' method. The 'appsettings.json' file has been updated for copy to the output directory.