Add [Key] attribute to model properties
The [Key] data annotation attribute has been added to primary keys in the Stack, StudySession, and Flashcard models. This is an improvement to the code because it provides a clearer definition for the Entity Framework, helping it understand the primary keys of these models better and hence improve the accuracy and efficiency of database operations.